image-rs: match hasher algorithm with exact prefix
Matching hasher algorithm with ":", such as "sha256:" instead of
"sha256".
