+++ image-rs/src/lib.rs
@@ -10,6 +10,7 @@ pub mod bundle;
 pub mod config;
 pub mod decoder;
 pub mod decrypt;
+pub mod digest;
 pub mod image;
 pub mod meta_store;
 #[cfg(feature = "nydus")]
